feat: refactor global credentials handling into a separate widget
